spring-jdbc/src/test/java/walsh/teaching/spring/jdbc/SpringJdbcBasedBookDaoTest.java
package walsh.teaching.spring.jdbc;

import com.github.springtestdbunit.DbUnitTestExecutionListener;
import com.github.springtestdbunit.annotation.DatabaseSetup;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.TestExecutionListeners;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration("classpath:beans.xml")
@TestExecutionListeners({DependencyInjectionTestExecutionListener.class,
        DbUnitTestExecutionListener.class})
public class SpringJdbcBasedBookDaoTest {
    @Autowired
    private SpringJdbcBasedBookDao bookDao;

    @Test
    @DatabaseSetup("classpath:dataset.xml")
    public void testFindAll() {
        assertThat(bookDao.findAll().size(), is(2));
    }
}
